# luann/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TypeAgent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agents/{agent_id}/config": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Config",
        "description": "Retrieve the configuration for a specific agent.\n\nThis endpoint fetches the configuration details for a given agent, identified by the user and agent IDs.",
        "operationId": "get_agent_config_api_agents__agent_id__config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/rename": {
      "patch": {
        "tags": [
          "agents"
        ],
        "summary": "Update Agent Name",
        "description": "Updates the name of a specific agent.\n\nThis changes the name of the agent in the database but does NOT edit the agent's persona.",
        "operationId": "update_agent_name_api_agents__agent_id__rename_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}": {
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "Delete Agent",
        "description": "Delete an agent.",
        "operationId": "delete_agent_api_agents__agent_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentRenameRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name",
            "description": "New name for the agent."
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentRenameRequest"
      },
      "AgentStateModel": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The unique identifier of the agent."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the agent."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "The unique identifier of the user associated with the agent."
          },
          "type_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Agent",
            "description": "The type agent."
          },
          "created_at": {
            "type": "integer",
            "title": "Created At",
            "description": "The unix timestamp of when the agent was created."
          },
          "preset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Preset Id",
            "description": "The unique identifier of the user associated with the agent."
          },
          "persona_memory": {
            "type": "string",
            "title": "Persona Memory",
            "description": "The persona used by the agent."
          },
          "human_memory": {
            "type": "string",
            "title": "Human Memory",
            "description": "The human used by the agent."
          },
          "llm_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LLMConfigModel"
              }
            ],
            "description": "The LLM configuration used by the agent."
          },
          "embedding_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmbeddingConfigModel"
              }
            ],
            "description": "The embedding configuration used by the agent."
          },
          "state": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "The state of the agent."
          },
          "user_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Status",
            "description": "The user status."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "user_id",
          "type_agent",
          "created_at",
          "preset_id",
          "persona_memory",
          "human_memory",
          "llm_config",
          "embedding_config",
          "user_status"
        ],
        "title": "AgentStateModel"
      },
...
      "EmbeddingConfigModel": {
        "properties": {
          "embedding_endpoint_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Endpoint Type",
            "default": "openai"
          },
          "embedding_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Endpoint",
            "default": "https://api.openai.com/v1"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model",
            "default": "text-embedding-ada-002"
          },
          "embedding_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dim",
            "default": 1536
          },
          "embedding_chunk_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Chunk Size",
            "default": 300
          }
        },
        "type": "object",
        "title": "EmbeddingConfigModel"
      },
...
      "GetAgentResponse": {
        "properties": {
          "agent_state": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AgentStateModel"
              }
            ],
            "description": "The state of the agent."
          }
        },
        "type": "object",
        "required": [
          "agent_state"
        ],
        "title": "GetAgentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMConfigModel": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "default": "gpt-4"
          },
          "model_endpoint_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Endpoint Type",
            "default": "openai"
          },
          "model_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Endpoint",
            "default": "https://api.openai.com/v1"
          },
          "model_wrapper": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Wrapper"
          },
          "context_window": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          }
        },
        "type": "object",
        "title": "LLMConfigModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://typeagent.localhost"
    },
    {
      "url": "http://localhost:8283"
    },
    {
      "url": "http://localhost:8083"
    }
  ]
}